+++ jingle_sleigh/src/ffi/image.rs
@@ -2,10 +2,8 @@ use crate::context::image::ImageProvider;
 use crate::ffi::instruction::bridge::VarnodeInfoFFI;
 use crate::VarNode;
 use cxx::ExternType;
+use crate::ffi::context_ffi::ImageFFI;
 
-pub(crate) struct ImageFFI<'a> {
-    provider: Box<dyn ImageProvider + 'a>,
-}
 
 impl<'a> ImageFFI<'a> {
     pub(crate) fn new<T: ImageProvider + 'a>(provider: T) -> Self {
@@ -13,7 +11,7 @@ impl<'a> ImageFFI<'a> {
             provider: Box::new(provider),
         }
     }
-    fn load(&self, vn: &VarnodeInfoFFI, out: &mut [u8]) -> usize {
+    pub(crate) fn load(&self, vn: &VarnodeInfoFFI, out: &mut [u8]) -> usize {
         self.provider.load(&VarNode::from(vn), out)
     }
 
@@ -25,16 +23,4 @@ impl<'a> ImageFFI<'a> {
 unsafe impl<'a> ExternType for ImageFFI<'a> {
     type Id = cxx::type_id!("ImageFFI");
     type Kind = cxx::kind::Opaque;
-}
-
-#[cxx::bridge]
-pub(crate) mod bridge {
-    unsafe extern "C++" {
-        type VarnodeInfoFFI = crate::ffi::instruction::bridge::VarnodeInfoFFI;
-    }
-    extern "Rust" {
-        include!("jingle_sleigh/src/ffi/instruction.rs.h");
-        type ImageFFI<'a>;
-        fn load(self: &ImageFFI, vn: &VarnodeInfoFFI, out: &mut [u8]) -> usize;
-    }
 }
